requirements.txt
langchain
openai
google-search-results
python-dotenv
